Fix one small issue with the noters if preparation beats set to 1 and added the kinematics of victory and defeat on the last level
